device-posture/device-posture-media-queries.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
'use strict';

// REFERENCE https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/http/tests/inspector-protocol/device-posture/device-posture-media-queries.js
promise_test(async (t) => {
  assert_equals(navigator.devicePosture.type, 'continuous');
  assert_true(matchMedia('(device-posture: continuous)').matches);

  const foldedMQL = window.matchMedia('(device-posture: folded)');
  const promise = new Promise(resolve => {
    foldedMQL.addEventListener(
      'change',
      () => { resolve(foldedMQL.matches); },
      { once: true }
    );
  });
  await test_driver.set_device_posture('folded');
  assert_true(await promise)
}, 'Device posture test: device-posture-media-queries.js.');
</script>
